pyproject.toml
[project]
name = "claude_vestige"
version = "0.1.0"
description = "Persistent semantic memory plugin for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.4.0",
    "fastembed>=0.2.0",
    "rank-bm25>=0.2.2",
    "pathspec>=0.11.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
mcp = [
    "mcp>=1.0.0",
]
dashboard = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["claude_vestige*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
